requirements.txt
# Django関連
Django==4.2.16
asgiref==3.9.1
sqlparse==0.5.3

# サーバー・静的ファイル
gunicorn==20.1.0
whitenoise==6.6.0

# データベース
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# OCR機能
Pillow==11.3.0
pyocr==0.8.5
opencv-python==4.12.0.88
numpy==2.2.6

# PDF生成機能
pdfkit==1.0.0
weasyprint==66.0
requests==2.32.4

# WeasyPrint依存関係
cssselect2==0.8.0
fonttools==4.59.0
pydyf==0.11.0
pyphen==0.17.2
tinycss2==1.4.0
tinyhtml5==2.0.0
webencodings==0.5.1
zopfli==0.2.3.post1
Brotli==1.1.0

# その他の依存関係
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
idna==3.10
pycparser==2.22
urllib3==2.5.0
